+++ NEWS.md
@@ -1,8 +1,11 @@
 # smoof 1.6
 
-## New Features
+## New features
 
 * Added bi-objective BBOB problem generator: makeBiObjBBOBFunction
+* Added: Kursawe function
+* Added: ED1 and ED2 functions
+* Added: WFG1 to WFG9 problems from the WFG-Problem Generator Toolkit
 
 # smoof 1.5.1
 
@@ -18,7 +21,7 @@
 * Better handling of additional arguments in makeFunctionsByName
 * convertTo{Minimization,Maximization} now flips sign of optima
 
-## Miscelaneous
+## Miscellaneous
 
 * logging and couting wrapper now have class smoof_function